Create ArraySampleKey so our keys are made up of the number of bytes, POD type, and MD5 digest.
The POD and extent are no longer added to the MD5 digest.

Fixed a bug where the MD5 digest for strings or wstrings was not taking into
account the extent to determine the total number of strings.

Fixed a bug for wstrings where even though we write them out as 32 bit, when
they are 16 bit the MD5 digest is calculated as if we are writing them out
as 16 bit instead of 32 bit.