Fixed the way the frame range is calculated, and store the start and end time on the node for users to query.
If a transform only has 1 sample for it's animation, treat it like it is static,
so we only set it once and don't create the connections for it.

Removed the useless timeRange command argument, and got fitTimeRange to actually
set the timeline based on the input file.

Split sequence attribute on node into two attributes, startFrame and endFrame.