Use MComputation so users can abort the export (by pressing the escape key, etc)
Look for faceVaryingInterpolateBoundary not facevaryingType when looking for
that SubD setting.

Write maya point and vector attributes as Alembic point, vector or normal
array properties based on whether a PROPNAME_AbcType attr exists.

PROPNAME_AbcGeomType now supports all the scopes of AbcGeomParam (defaults to
kConstant) and uses the same short string to indicate it.

Fixed the help message to mention -a instead of -at for -attr.
Added a warning about how exporting could fail because of sibling nodes with
the same name if the -sn (strip namespace) flag was used.

Improved writing of width data on curves.