Grab some of Brians ArchiveInfo related changes and move them around.
Keep the ArchiveInfo stuff in Abc (instead of AbcCoreAbstract)
Instead of overriding the MetaData, have a helper function that allows users
to create the archive like the OArchive constructor but with additional data.

Store only one API version number in Foundation, and add a test to
AbcCoreAbstract to make sure that the version is numeric and within a certain
range.