Change the internal file version in anticipation of major changes to improve walking the hierarchy.
Take advantage of some unused bits to store more compact forms of the most
common types of meta data. (geoScope, interpretation, and schema)

This helps to save a little disk space, but this change by itself hurts tree
walking performance. (we'll get better results with stuff we can do
BECAUSE of this change)