Remove .prop and .geom/.xform/.whatever HDF5 groups.
Schema Objects now use the "top" compound property as the location for it's
properties.

Since .prop was removed, we needed a way to distinguish between an HDF5 group
which is meant to be a child Object, and what is meant to be a Compound
Property (at a given object).  We currently do this by writing out a string of
properties for  the list of child compounds, and a list of properties for the
child objects.

We still need to measure if writing this extra data is faster or slower than
H5Aiterate and H5Literate.

Changing this relationship meant we have to deal with Issue 171
Scalar and Array Properties can run into trouble when parented to the top
compound of an OObject.

To fix this bug, we used composition instead of inheritance, so the
BaseCpwImpl, TopCpwImpl, BaseOwImpl, and TopImpl are no longer necessary.

Instead we introduce CpwData and OwData to hold a bunch of this information.

Similiar changes will need to be made on the read side.