Adding ILight/OLight.
Some interesting things to note:

The light is made up of all optional properties.

Since everything is optional, whether or not a light is constant is
determined by the child bounds property.

The only way to write data to the optional camera schema is via the
setCameraSample command.  Since CameraSample can have child bounds data, but
child bounds doesn't make sense on the optional camera so it gets stripped.

Important data regarding the light will live in the upcoming I/OMaterial schema.