Remove kWrapExisting constructor from OObject since it doesn't work well because the properties an O*SchemaObject may hold can go out of scope when downcasting to OObject, AND O*SchemaObject may also contain some state that isn't transferable.
Instead make OObject polymorphic by making the destructor virtual.
It needs to be polymorphic so we can wrap it in a shared_ptr (which we
add convenient typedefs for).

A user can now get a downcasted OObjectPtr and use dynamic_pointer_cast to
get the appropriate O*Ptr.

We also make the destructor for IObject virtual and create convenient typedefs
so that technique works on the ISchemaObjectas well. (kWrapExisting still exists
on IObject)