Add a few tests for iCacheHierarchy on AbcCoreHDF5 Read/Write.
Change the behavior of the iCacheHierarchy flag on read.

If the extra hierarchy data exists read it if the flag is set to true, if
the data doesn't exist, or the flag is set to false, traverse the hierarchy
like usual.

Fix a bug where .info wasn't being removed when dealing with the extra cached
hierarchy.