A bunch of fixes to get the Array tests working properly.
We now write the dimensions after the data since the data could throw
because of a null check on strings.

Other fixes include throwing when bad pod types or incorrect property types
are set, better handling of empty data for dimensions (and returning early)
and writing out the key for empty strings.

Currently strings share from a different buffer in the array sample map, this
is because of the special packing logic we do, and how calculating the number
of bytes from the pod doesn't currently work for strings since we don't loop
through each one.