Fix crash in AbcLs when providing an index to a compound property.
Fix bug where not all values in the array sample were being printed out and the
formatting for the first and last value was off when the extent was not 1.

Add flag "-s" for printing just the size of a particular array property sample.
For a scalar it just prints the extent.

Add flags "-t" and "-f" for printing out time data on the archive and what
time a particular index sample corresponds with.
-t is time in seconds, -f is the frame time where fps is currently assumed to
be 24.